Refactor sole application AoS notifications
