+++ src/main/java/uk/gov/hmcts/divorce/citizen/notification/SoleApplicationDisputedNotification.java
@@ -1,0 +1,58 @@
+package uk.gov.hmcts.divorce.citizen.notification;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+import uk.gov.hmcts.divorce.divorcecase.model.Applicant;
+import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
+import uk.gov.hmcts.divorce.notification.ApplicantNotification;
+import uk.gov.hmcts.divorce.notification.CommonContent;
+import uk.gov.hmcts.divorce.notification.NotificationService;
+
+import java.util.Map;
+
+import static uk.gov.hmcts.divorce.notification.CommonContent.SUBMISSION_RESPONSE_DATE;
+import static uk.gov.hmcts.divorce.notification.EmailTemplateName.SOLE_APPLICANT_DISPUTED_AOS_SUBMITTED;
+import static uk.gov.hmcts.divorce.notification.EmailTemplateName.SOLE_RESPONDENT_DISPUTED_AOS_SUBMITTED;
+import static uk.gov.hmcts.divorce.notification.FormatUtil.DATE_TIME_FORMATTER;
+
+@Component
+@Slf4j
+public class SoleApplicationDisputedNotification implements ApplicantNotification {
+
+    @Autowired
+    private NotificationService notificationService;
+
+    @Autowired
+    private CommonContent commonContent;
+
+    @Override
+    public void sendToApplicant1(final CaseData caseData, final Long id) {
+        log.info("Sending Aos disputed notification to applicant");
+
+        notificationService.sendEmail(
+            caseData.getApplicant1().getEmail(),
+            SOLE_APPLICANT_DISPUTED_AOS_SUBMITTED,
+            disputedTemplateVars(caseData, id, caseData.getApplicant1(), caseData.getApplicant2()),
+            caseData.getApplicant1().getLanguagePreference()
+        );
+    }
+
+    @Override
+    public void sendToApplicant2(final CaseData caseData, final Long id) {
+        log.info("Sending Aos submitted disputed notification to respondent");
+
+        notificationService.sendEmail(
+            caseData.getApplicant2EmailAddress(),
+            SOLE_RESPONDENT_DISPUTED_AOS_SUBMITTED,
+            disputedTemplateVars(caseData, id, caseData.getApplicant2(), caseData.getApplicant1()),
+            caseData.getApplicant2().getLanguagePreference()
+        );
+    }
+
+    private Map<String, String> disputedTemplateVars(CaseData caseData, Long id, Applicant applicant, Applicant partner) {
+        Map<String, String> templateVars = commonContent.mainTemplateVars(caseData, id, applicant, partner);
+        templateVars.put(SUBMISSION_RESPONSE_DATE, caseData.getDueDate().format(DATE_TIME_FORMATTER));
+        return templateVars;
+    }
+}
