+++ src/main/java/uk/gov/hmcts/divorce/citizen/notification/SoleApplicationNotDisputedNotification.java
@@ -5,21 +5,19 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.divorce.divorcecase.model.Applicant;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
+import uk.gov.hmcts.divorce.notification.ApplicantNotification;
 import uk.gov.hmcts.divorce.notification.CommonContent;
 import uk.gov.hmcts.divorce.notification.NotificationService;
 
 import java.util.Map;
 
-import static uk.gov.hmcts.divorce.notification.CommonContent.SUBMISSION_RESPONSE_DATE;
 import static uk.gov.hmcts.divorce.notification.EmailTemplateName.SOLE_APPLICANT_AOS_SUBMITTED;
-import static uk.gov.hmcts.divorce.notification.EmailTemplateName.SOLE_APPLICANT_DISPUTED_AOS_SUBMITTED;
 import static uk.gov.hmcts.divorce.notification.EmailTemplateName.SOLE_RESPONDENT_AOS_SUBMITTED;
-import static uk.gov.hmcts.divorce.notification.EmailTemplateName.SOLE_RESPONDENT_DISPUTED_AOS_SUBMITTED;
 import static uk.gov.hmcts.divorce.notification.FormatUtil.DATE_TIME_FORMATTER;
 
 @Component
 @Slf4j
-public class SoleAosSubmittedNotification {
+public class SoleApplicationNotDisputedNotification implements ApplicantNotification {
 
     private static final String APPLY_FOR_CO_DATE = "apply for CO date";
 
@@ -29,7 +27,8 @@ public class SoleAosSubmittedNotification {
     @Autowired
     private CommonContent commonContent;
 
-    public void sendApplicationNotDisputedToApplicant(CaseData caseData, Long id) {
+    @Override
+    public void sendToApplicant1(final CaseData caseData, final Long id) {
         log.info("Sending Aos submitted without dispute notification to applicant");
 
         notificationService.sendEmail(
@@ -40,7 +39,8 @@ public class SoleAosSubmittedNotification {
         );
     }
 
-    public void sendApplicationNotDisputedToRespondent(CaseData caseData, Long id) {
+    @Override
+    public void sendToApplicant2(CaseData caseData, Long id) {
         log.info("Sending Aos submitted without dispute notification to respondent");
 
         notificationService.sendEmail(
@@ -51,37 +51,9 @@ public class SoleAosSubmittedNotification {
         );
     }
 
-    public void sendApplicationDisputedToApplicant(CaseData caseData, Long id) {
-        log.info("Sending Aos disputed notification to applicant");
-
-        notificationService.sendEmail(
-            caseData.getApplicant1().getEmail(),
-            SOLE_APPLICANT_DISPUTED_AOS_SUBMITTED,
-            disputedTemplateVars(caseData, id, caseData.getApplicant1(), caseData.getApplicant2()),
-            caseData.getApplicant1().getLanguagePreference()
-        );
-    }
-
-    public void sendApplicationDisputedToRespondent(CaseData caseData, Long id) {
-        log.info("Sending Aos submitted disputed notification to respondent");
-
-        notificationService.sendEmail(
-            caseData.getApplicant2EmailAddress(),
-            SOLE_RESPONDENT_DISPUTED_AOS_SUBMITTED,
-            disputedTemplateVars(caseData, id, caseData.getApplicant2(), caseData.getApplicant1()),
-            caseData.getApplicant2().getLanguagePreference()
-        );
-    }
-
     private Map<String, String> nonDdisputedTemplateVars(CaseData caseData, Long id, Applicant applicant, Applicant partner) {
         Map<String, String> templateVars = commonContent.mainTemplateVars(caseData, id, applicant, partner);
         templateVars.put(APPLY_FOR_CO_DATE, caseData.getDueDate().format(DATE_TIME_FORMATTER));
-        return templateVars;
-    }
-
-    private Map<String, String> disputedTemplateVars(CaseData caseData, Long id, Applicant applicant, Applicant partner) {
-        Map<String, String> templateVars = commonContent.mainTemplateVars(caseData, id, applicant, partner);
-        templateVars.put(SUBMISSION_RESPONSE_DATE, caseData.getDueDate().format(DATE_TIME_FORMATTER));
         return templateVars;
     }
 }
