+++ src/main/java/uk/gov/hmcts/divorce/common/service/task/SendCitizenAosNotifications.java
@@ -3,35 +3,33 @@ package uk.gov.hmcts.divorce.common.service.task;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.ccd.sdk.api.CaseDetails;
-import uk.gov.hmcts.divorce.citizen.notification.SoleAosSubmittedNotification;
+import uk.gov.hmcts.divorce.citizen.notification.SoleApplicationDisputedNotification;
+import uk.gov.hmcts.divorce.citizen.notification.SoleApplicationNotDisputedNotification;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
 import uk.gov.hmcts.divorce.divorcecase.model.State;
 import uk.gov.hmcts.divorce.divorcecase.task.CaseTask;
+import uk.gov.hmcts.divorce.notification.NotificationDispatcher;
 
 @Component
 public class SendCitizenAosNotifications implements CaseTask {
 
     @Autowired
-    private SoleAosSubmittedNotification soleAosSubmittedNotification;
+    private SoleApplicationNotDisputedNotification soleApplicationNotDisputedNotification;
+
+    @Autowired
+    private SoleApplicationDisputedNotification soleApplicationDisputedNotification;
+
+    @Autowired
+    private NotificationDispatcher notificationDispatcher;
 
     @Override
     public CaseDetails<CaseData, State> apply(CaseDetails<CaseData, State> details) {
         final var data = details.getData();
 
         if (data.getAcknowledgementOfService().isDisputed()) {
-            if (!data.getApplicant1().isRepresented()) {
-                soleAosSubmittedNotification.sendApplicationDisputedToApplicant(data, details.getId());
-            }
-            if (!data.getApplicant2().isRepresented()) {
-                soleAosSubmittedNotification.sendApplicationDisputedToRespondent(data, details.getId());
-            }
+            notificationDispatcher.send(soleApplicationDisputedNotification, data, details.getId());
         } else {
-            if (!data.getApplicant1().isRepresented()) {
-                soleAosSubmittedNotification.sendApplicationNotDisputedToApplicant(data, details.getId());
-            }
-            if (!data.getApplicant2().isRepresented()) {
-                soleAosSubmittedNotification.sendApplicationNotDisputedToRespondent(data, details.getId());
-            }
+            notificationDispatcher.send(soleApplicationNotDisputedNotification, data, details.getId());
         }
 
         return details;
