+++ src/test/java/uk/gov/hmcts/divorce/citizen/notification/SoleApplicationDisputedNotificationTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
+import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
 import uk.gov.hmcts.divorce.notification.CommonContent;
 import uk.gov.hmcts.divorce.notification.NotificationService;
@@ -26,9 +26,7 @@ import static uk.gov.hmcts.divorce.notification.CommonContent.IS_DIVORCE;
 import static uk.gov.hmcts.divorce.notification.CommonContent.NO;
 import static uk.gov.hmcts.divorce.notification.CommonContent.SUBMISSION_RESPONSE_DATE;
 import static uk.gov.hmcts.divorce.notification.CommonContent.YES;
-import static uk.gov.hmcts.divorce.notification.EmailTemplateName.SOLE_APPLICANT_AOS_SUBMITTED;
 import static uk.gov.hmcts.divorce.notification.EmailTemplateName.SOLE_APPLICANT_DISPUTED_AOS_SUBMITTED;
-import static uk.gov.hmcts.divorce.notification.EmailTemplateName.SOLE_RESPONDENT_AOS_SUBMITTED;
 import static uk.gov.hmcts.divorce.notification.EmailTemplateName.SOLE_RESPONDENT_DISPUTED_AOS_SUBMITTED;
 import static uk.gov.hmcts.divorce.notification.FormatUtil.DATE_TIME_FORMATTER;
 import static uk.gov.hmcts.divorce.notification.FormatUtil.formatId;
@@ -37,8 +35,8 @@ import static uk.gov.hmcts.divorce.testutil.TestConstants.TEST_USER_EMAIL;
 import static uk.gov.hmcts.divorce.testutil.TestDataHelper.getMainTemplateVars;
 import static uk.gov.hmcts.divorce.testutil.TestDataHelper.validCaseDataForAosSubmitted;
 
-@ExtendWith(SpringExtension.class)
-public class SoleAosSubmittedNotificationTest {
+@ExtendWith(MockitoExtension.class)
+class SoleApplicationDisputedNotificationTest {
 
     @Mock
     private NotificationService notificationService;
@@ -47,107 +45,7 @@ public class SoleAosSubmittedNotificationTest {
     private CommonContent commonContent;
 
     @InjectMocks
-    private SoleAosSubmittedNotification notification;
-
-    @Test
-    void shouldSendAosNotDisputedEmailToSoleApplicantWithDivorceContent() {
-        CaseData data = validCaseDataForAosSubmitted();
-        data.setDueDate(LocalDate.now().plusDays(141));
-        when(commonContent.mainTemplateVars(data, 1234567890123456L, data.getApplicant1(), data.getApplicant2()))
-            .thenReturn(getMainTemplateVars());
-
-        notification.sendApplicationNotDisputedToApplicant(data, 1234567890123456L);
-
-        verify(notificationService).sendEmail(
-            eq(TEST_USER_EMAIL),
-            eq(SOLE_APPLICANT_AOS_SUBMITTED),
-            argThat(allOf(
-                hasEntry(APPLICATION_REFERENCE, formatId(1234567890123456L)),
-                hasEntry("apply for CO date", data.getDueDate().format(DATE_TIME_FORMATTER)),
-                hasEntry(IS_DIVORCE, YES),
-                hasEntry(IS_DISSOLUTION, NO)
-            )),
-            eq(ENGLISH)
-        );
-        verify(commonContent).mainTemplateVars(data, 1234567890123456L, data.getApplicant1(), data.getApplicant2());
-    }
-
-    @Test
-    void shouldSendAosNotDisputedEmailToSoleApplicantWithDissolutionContent() {
-        CaseData data = validCaseDataForAosSubmitted();
-        data.setDivorceOrDissolution(DISSOLUTION);
-        data.setDueDate(LocalDate.now().plusDays(141));
-        final Map<String, String> templateVars = getMainTemplateVars();
-        templateVars.putAll(Map.of(IS_DISSOLUTION, YES, IS_DIVORCE, NO));
-        when(commonContent.mainTemplateVars(data, 1234567890123456L, data.getApplicant1(), data.getApplicant2())).thenReturn(templateVars);
-
-        notification.sendApplicationNotDisputedToApplicant(data, 1234567890123456L);
-
-        verify(notificationService).sendEmail(
-            eq(TEST_USER_EMAIL),
-            eq(SOLE_APPLICANT_AOS_SUBMITTED),
-            argThat(allOf(
-                hasEntry(APPLICATION_REFERENCE, formatId(1234567890123456L)),
-                hasEntry("apply for CO date", data.getDueDate().format(DATE_TIME_FORMATTER)),
-                hasEntry(IS_DIVORCE, NO),
-                hasEntry(IS_DISSOLUTION, YES)
-            )),
-            eq(ENGLISH)
-        );
-        verify(commonContent).mainTemplateVars(data, 1234567890123456L, data.getApplicant1(), data.getApplicant2());
-    }
-
-    @Test
-    void shouldSendAosNotDisputedEmailToSoleRespondentWithDivorceContent() {
-        CaseData data = validCaseDataForAosSubmitted();
-        data.setDueDate(LocalDate.now().plusDays(141));
-        data.getApplicant2().setEmail(null);
-
-        when(commonContent.mainTemplateVars(data, 1234567890123456L, data.getApplicant2(), data.getApplicant1()))
-            .thenReturn(getMainTemplateVars());
-
-        notification.sendApplicationNotDisputedToRespondent(data, 1234567890123456L);
-
-        verify(notificationService).sendEmail(
-            eq(TEST_APPLICANT_2_USER_EMAIL),
-            eq(SOLE_RESPONDENT_AOS_SUBMITTED),
-            argThat(allOf(
-                hasEntry(APPLICATION_REFERENCE, formatId(1234567890123456L)),
-                hasEntry("apply for CO date", data.getDueDate().format(DATE_TIME_FORMATTER)),
-                hasEntry(IS_DIVORCE, YES),
-                hasEntry(IS_DISSOLUTION, NO)
-            )),
-            eq(ENGLISH)
-        );
-        verify(commonContent).mainTemplateVars(data, 1234567890123456L, data.getApplicant2(), data.getApplicant1());
-    }
-
-    @Test
-    void shouldSendAosNotDisputedEmailToSoleRespondentWithDissolutionContent() {
-        CaseData data = validCaseDataForAosSubmitted();
-        data.setDivorceOrDissolution(DISSOLUTION);
-        data.setDueDate(LocalDate.now().plusDays(141));
-        data.getApplicant2().setEmail(null);
-
-        final Map<String, String> templateVars = getMainTemplateVars();
-        templateVars.putAll(Map.of(IS_DISSOLUTION, YES, IS_DIVORCE, NO));
-        when(commonContent.mainTemplateVars(data, 1234567890123456L, data.getApplicant2(), data.getApplicant1())).thenReturn(templateVars);
-
-        notification.sendApplicationNotDisputedToRespondent(data, 1234567890123456L);
-
-        verify(notificationService).sendEmail(
-            eq(TEST_APPLICANT_2_USER_EMAIL),
-            eq(SOLE_RESPONDENT_AOS_SUBMITTED),
-            argThat(allOf(
-                hasEntry(APPLICATION_REFERENCE, formatId(1234567890123456L)),
-                hasEntry("apply for CO date", data.getDueDate().format(DATE_TIME_FORMATTER)),
-                hasEntry(IS_DIVORCE, NO),
-                hasEntry(IS_DISSOLUTION, YES)
-            )),
-            eq(ENGLISH)
-        );
-        verify(commonContent).mainTemplateVars(data, 1234567890123456L, data.getApplicant2(), data.getApplicant1());
-    }
+    private SoleApplicationDisputedNotification soleApplicationDisputedNotification;
 
     @Test
     void shouldSendAosDisputedEmailToSoleApplicantWithDivorceContent() {
@@ -157,7 +55,7 @@ public class SoleAosSubmittedNotificationTest {
         when(commonContent.mainTemplateVars(data, 1234567890123456L, data.getApplicant1(), data.getApplicant2()))
             .thenReturn(getMainTemplateVars());
 
-        notification.sendApplicationDisputedToApplicant(data, 1234567890123456L);
+        soleApplicationDisputedNotification.sendToApplicant1(data, 1234567890123456L);
 
         verify(notificationService).sendEmail(
             eq(TEST_USER_EMAIL),
@@ -183,7 +81,7 @@ public class SoleAosSubmittedNotificationTest {
         templateVars.putAll(Map.of(IS_DISSOLUTION, YES, IS_DIVORCE, NO));
         when(commonContent.mainTemplateVars(data, 1234567890123456L, data.getApplicant1(), data.getApplicant2())).thenReturn(templateVars);
 
-        notification.sendApplicationDisputedToApplicant(data, 1234567890123456L);
+        soleApplicationDisputedNotification.sendToApplicant1(data, 1234567890123456L);
 
         verify(notificationService).sendEmail(
             eq(TEST_USER_EMAIL),
@@ -209,7 +107,7 @@ public class SoleAosSubmittedNotificationTest {
         when(commonContent.mainTemplateVars(data, 1234567890123456L, data.getApplicant2(), data.getApplicant1()))
             .thenReturn(getMainTemplateVars());
 
-        notification.sendApplicationDisputedToRespondent(data, 1234567890123456L);
+        soleApplicationDisputedNotification.sendToApplicant2(data, 1234567890123456L);
 
         verify(notificationService).sendEmail(
             eq(TEST_APPLICANT_2_USER_EMAIL),
@@ -237,7 +135,7 @@ public class SoleAosSubmittedNotificationTest {
         templateVars.putAll(Map.of(IS_DISSOLUTION, YES, IS_DIVORCE, NO));
         when(commonContent.mainTemplateVars(data, 1234567890123456L, data.getApplicant2(), data.getApplicant1())).thenReturn(templateVars);
 
-        notification.sendApplicationDisputedToRespondent(data, 1234567890123456L);
+        soleApplicationDisputedNotification.sendToApplicant2(data, 1234567890123456L);
 
         verify(notificationService).sendEmail(
             eq(TEST_APPLICANT_2_USER_EMAIL),
