+++ src/test/java/uk/gov/hmcts/divorce/common/service/task/SendCitizenAosNotificationsTest.java
@@ -1,0 +1,73 @@
+package uk.gov.hmcts.divorce.common.service.task;
+
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+import uk.gov.hmcts.ccd.sdk.api.CaseDetails;
+import uk.gov.hmcts.divorce.citizen.notification.SoleApplicationDisputedNotification;
+import uk.gov.hmcts.divorce.citizen.notification.SoleApplicationNotDisputedNotification;
+import uk.gov.hmcts.divorce.divorcecase.model.AcknowledgementOfService;
+import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
+import uk.gov.hmcts.divorce.divorcecase.model.State;
+import uk.gov.hmcts.divorce.notification.NotificationDispatcher;
+
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyNoMoreInteractions;
+import static uk.gov.hmcts.divorce.divorcecase.model.HowToRespondApplication.DISPUTE_DIVORCE;
+import static uk.gov.hmcts.divorce.divorcecase.model.HowToRespondApplication.WITHOUT_DISPUTE_DIVORCE;
+
+@ExtendWith(MockitoExtension.class)
+class SendCitizenAosNotificationsTest {
+
+    @Mock
+    private SoleApplicationNotDisputedNotification soleApplicationNotDisputedNotification;
+
+    @Mock
+    private SoleApplicationDisputedNotification soleApplicationDisputedNotification;
+
+    @Mock
+    private NotificationDispatcher notificationDispatcher;
+
+    @InjectMocks
+    private SendCitizenAosNotifications sendCitizenAosNotifications;
+
+    @Test
+    void shouldSendDisputedNotifications() {
+
+        final CaseData caseData = CaseData.builder()
+            .acknowledgementOfService(AcknowledgementOfService.builder()
+                .howToRespondApplication(DISPUTE_DIVORCE)
+                .build())
+            .build();
+
+        final CaseDetails<CaseData, State> caseDetails = new CaseDetails<>();
+        caseDetails.setId(1L);
+        caseDetails.setData(caseData);
+
+        sendCitizenAosNotifications.apply(caseDetails);
+
+        verify(notificationDispatcher).send(soleApplicationDisputedNotification, caseData, 1L);
+        verifyNoMoreInteractions(notificationDispatcher);
+    }
+
+    @Test
+    void shouldSendNotDisputedNotifications() {
+
+        final CaseData caseData = CaseData.builder()
+            .acknowledgementOfService(AcknowledgementOfService.builder()
+                .howToRespondApplication(WITHOUT_DISPUTE_DIVORCE)
+                .build())
+            .build();
+
+        final CaseDetails<CaseData, State> caseDetails = new CaseDetails<>();
+        caseDetails.setId(1L);
+        caseDetails.setData(caseData);
+
+        sendCitizenAosNotifications.apply(caseDetails);
+
+        verify(notificationDispatcher).send(soleApplicationNotDisputedNotification, caseData, 1L);
+        verifyNoMoreInteractions(notificationDispatcher);
+    }
+}
